rename cross component to XMark and update imports
