+++ components/xmark/XMark.stories.tsx
@@ -1,12 +1,12 @@
 import { Meta, StoryFn } from "@storybook/react";
-import Cross from "./Cross";
+import XMark from "./XMark";
 
 export default {
   title: "Cross",
-  component: Cross,
-} as Meta<typeof Cross>;
+  component: XMark,
+} as Meta<typeof XMark>;
 
-const Template: StoryFn<typeof Cross> = (args) => <Cross {...args} />;
+const Template: StoryFn<typeof XMark> = (args) => <XMark {...args} />;
 
 export const BlackCross = Template.bind({});
 BlackCross.args = { className: "" };
